Refactors the tab activities into fragments
