--- app/src/main/res/layout/activity_movie.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="horizontal"
-    android:weightSum="5"
-    tools:context=".MovieActivity"
-    tools:deviceIds="tv"
-    tools:ignore="MergeRootFrame"
-    android:baselineAligned="false">
-
-    <FrameLayout
-        android:id="@+id/movie_fragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_margin="5dp"
-        android:baselineAligned="false"
-        android:layout_weight="3" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_weight="2"
-        android:orientation="vertical">
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:textSize="20sp"
-            android:text="@string/stashapp_scenes" />
-
-        <FrameLayout
-            android:id="@+id/movie_list_fragment"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_margin="5dp" />
-    </LinearLayout>
-
-</LinearLayout>
