--- app/src/main/res/layout/gallery_activity.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <com.github.damontecres.stashapp.views.TabbedGridTitleView
-        android:id="@id/browse_title_group"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <com.github.damontecres.stashapp.views.HomeImageButton
-            android:id="@+id/icon"
-            android:layout_width="@dimen/title_bar_icon_size"
-            android:layout_height="@dimen/title_bar_icon_size"
-            android:layout_margin="@dimen/title_bar_margin"
-            android:contentDescription="@string/stashapp_image"
-            android:layout_alignParentStart="true"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/grid_title"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_marginStart="64dp"
-            android:paddingEnd="0dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="Test Title"
-            style="@style/GridTitleText" />
-
-    </com.github.damontecres.stashapp.views.TabbedGridTitleView>
-
-    <androidx.leanback.tab.LeanbackTabLayout
-        android:id="@+id/tab_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:tabGravity="center"
-        app:tabIndicatorColor="@color/selected_background"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.leanback.tab.LeanbackViewPager
-        android:id="@+id/view_pager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:deviceIds="tv"
-        tools:ignore="MergeRootFrame"
-        />
-</LinearLayout>
